Simplification, cross-compatibility and corrections
The separator "/" was changed to filesep, which allows matlab path to work for windows or unix systems.
The Matpro function was unified into a single one in the 00.Lib folder. All exercises that need matpro, call to this one. Same was done for H2Oeq.m
Normalization of flux was corrected for volume of material, so flux in cladding is in between fuel and coolant.
Slight improvement to plotting of results and naming.